refactor: FormCollapsible 로 변경, 스타일 수정
