+++ package.json
@@ -30,8 +30,7 @@
     "@emotion/styled": "^11.8.1",
     "@headlessui/react": "^1.6.6",
     "@hookform/resolvers": "^2.9.1",
-    "@radix-ui/react-accordion": "^1.1.1",
-    "@radix-ui/react-collapsible": "^1.0.1",
+    "@radix-ui/react-collapsible": "^1.0.2",
     "@radix-ui/react-dialog": "^1.0.2",
     "@radix-ui/react-dropdown-menu": "^2.0.2",
     "@radix-ui/react-progress": "^1.0.1",
